+++ src/map/MapPlot.cpp
@@ -27,7 +27,8 @@ MapPlot::MapPlot() : _impl{std::make_unique<Impl>()}, _loader{std::make_shared<T
 }
 
 MapPlot::MapPlot(std::shared_ptr<ITileLoader>& loader)
-    : _impl{std::make_unique<Impl>()}, _loader{loader}, _minLat{MIN_LAT}, _minLon{MIN_LON}, _maxLat{MAX_LAT}, _maxLon{MAX_LON} {
+    : _impl{std::make_unique<Impl>()}, _loader{loader}, _minLat{(float) MIN_LAT}, _minLon{(float) MIN_LON}, _maxLat{(float) MAX_LAT},
+      _maxLon{(float) MAX_LON} {
     resetBounds();
 }
 
@@ -41,10 +42,10 @@ void MapPlot::paint() {
 
         if (_setBounds != SetBounds::None) {
             if (_setBounds == SetBounds::Geo) {
-                _minX = lon2x(_minLon, 0);
-                _maxX = lon2x(_maxLon, 0);
-                _minY = lat2y(_minLat, 0);
-                _maxY = lat2y(_maxLat, 0);
+                _minX = (float) lon2x(_minLon, 0);
+                _maxX = (float) lon2x(_maxLon, 0);
+                _minY = (float) lat2y(_minLat, 0);
+                _maxY = (float) lat2y(_maxLat, 0);
             } else if (_setBounds == SetBounds::Local) {
                 // do nothing
             }
@@ -59,16 +60,16 @@ void MapPlot::paint() {
         _impl->plotLims = ImPlot::GetPlotLimits(ImAxis_X1, ImAxis_Y1);
         _impl->plotSize = ImPlot::GetPlotSize();
 
-        _mouseLon = x2lon(_impl->mousePos.x, 0);
-        _mouseLat = y2lat(_impl->mousePos.y, 0);
+        _mouseLon = (float) x2lon(_impl->mousePos.x, 0);
+        _mouseLat = (float) y2lat(_impl->mousePos.y, 0);
 
         _pixelsX = _impl->plotSize.x;
         _pixelsY = _impl->plotSize.y;
 
-        _minX = _impl->plotLims.X.Min;
-        _maxX = _impl->plotLims.X.Max;
-        _minY = _impl->plotLims.Y.Min;
-        _maxY = _impl->plotLims.Y.Max;
+        _minX = (float) _impl->plotLims.X.Min;
+        _maxX = (float) _impl->plotLims.X.Max;
+        _minY = (float) _impl->plotLims.Y.Min;
+        _maxY = (float) _impl->plotLims.Y.Max;
         _rangeX = fabs(_maxX - _minX);
         _rangeY = fabs(_maxY - _minY);
 
@@ -76,15 +77,15 @@ void MapPlot::paint() {
         _resY = _pixelsY / _rangeY;
         _zoom = std::clamp(int(floor(log2(_resX / _tilePixels))), MIN_ZOOM, MAX_ZOOM);
         _tilesNum = POW2[_zoom];
-        _tileSize = 1.0 / float(_tilesNum);
+        _tileSize = 1.0f / float(_tilesNum);
 
         const auto minMaxLat{std::minmax(y2lat(_minY * _tilesNum, _zoom), y2lat(_maxY * _tilesNum, _zoom))};
         const auto minMaxLon{std::minmax(x2lon(_minX * _tilesNum, _zoom), x2lon(_maxX * _tilesNum, _zoom))};
 
-        _minLat = minMaxLat.first;
-        _maxLat = minMaxLat.second;
-        _minLon = minMaxLon.first;
-        _maxLon = minMaxLon.second;
+        _minLat = (float) minMaxLat.first;
+        _maxLat = (float) minMaxLat.second;
+        _minLon = (float) minMaxLon.first;
+        _maxLon = (float) minMaxLon.second;
 
         _minTX = std::clamp(int(_minX * _tilesNum), 0, _tilesNum - 1);
         _maxTX = std::clamp(int(_maxX * _tilesNum), 0, _tilesNum - 1);
@@ -115,10 +116,10 @@ void MapPlot::paint() {
 }
 
 void MapPlot::resetBounds() {
-    _minLat = MIN_LAT;
-    _maxLat = MAX_LAT;
-    _minLon = MIN_LON;
-    _maxLon = MAX_LON;
+    _minLat = (float) MIN_LAT;
+    _maxLat = (float) MAX_LAT;
+    _minLon = (float) MIN_LON;
+    _maxLon = (float) MAX_LON;
     _setBounds = SetBounds::Geo;
 }
